+++ analytics/alliance_intelligence/__init__.py
@@ -1,0 +1,4 @@
+"""
+Sentinel
+Alliance Intelligence Package
+"""